Ein paar Änderungen für Windows. Jan
git-svn-id: svn+ssh://elmi.iti.uni-luebeck.de/export/svn/Hanse/branches/threadv1@881 8f13a6d9-6203-4c43-9b9f-a3039379152f

+++ Hanse/Framework/robotmodule.cpp
@@ -28,7 +28,7 @@ RobotModule::RobotModule(QString newId)
 
 void RobotModule::run()
 {
-    logger->info("Started thread id="+QString::number(QThread::currentThreadId()));
+    logger->info("Started thread id="+QString::number((int)QThread::currentThreadId()));
 
     // start health check timer
     healthCheckTimer.start();
